Unit test for item counter on homepage
