--- index.html
@@ -1,41 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Javascript Capstone</title>
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css">
-  </head>
-  <body>
-    <header>
-      <nav>
-        <div>
-          <a href="#" class="logo">
-            <div class="logo-image"></div>
-          </a>
-        </div>
-        <ul>
-          <li id="rockets-menu-item">
-            <a href="#">Rockets</a>
-          </li>
-          <li><a href="#">Planets</a></li>
-          <li><a href="#">Races</a></li>
-        </ul>
-      </nav>
-    </header>
-
-    <main>
-      <section id="home-container">
-        <ul class="rockets-container"></ul>
-      </section>
-      <section class="popups-container"></section>
-    </main>
-
-    <footer class="footer">
-      <p>Created by Dennis and Supreeti under MIT license</p>
-  </footer>
-    <script defer src="runtime.bundle.js"></script>
-    <script defer src="main.bundle.js"></script>
-  </body>
-</html>
