+++ src/index.html
@@ -12,7 +12,7 @@
       <nav>
         <div>
           <a href="#" class="logo">
-            <img src=" " alt="Your Logo" class="logo-image">
+            <div class="logo-image"></div>
           </a>
         </div>
         <ul>
@@ -27,7 +27,6 @@
 
     <main>
       <section id="home-container">
-        <!-- <i class="fas fa-times-circle" id='modal-btn'></i> -->
         <ul class="rockets-container"></ul>
       </section>
       <section class="popups-container"></section>
