+++ src/style.css
@@ -29,12 +29,17 @@ nav {
   text-decoration: none;
   color: #333;
   display: inline;
+  display: flex;
+  align-items: center; /* Align logo and text vertically */
 }
 
 .logo-image {
+  background-image: url('./assets/rocket-logo.svg');
+  background-size: contain;
+  background-repeat: no-repeat;
   width: 300px;
   height: 50px;
-  border-radius: 50%;
+  margin-right: 10px; /* Add some spacing between logo and text */
 }
 
 nav ul {
